# Add in running window optimization using scan