… integration_tests/src/main/python/window_function_test.py