@revans2 2021-07-09
I tested this on databricks and it works there too.