@revans2 2021-07-12
With the review work I accidentally checked in a change that makes this require the fix from https://github.com/rapidsai/cudf/pull/8705. I am inclined to wait for it to get merged in, but if others want to merge this in sooner I can revert the small change and do a follow on PR when the cudf change does get merged in.